Fix compile errors/warnings with gcc 4.3  - added missing #includes that were implicitly included via old headers.  - add namespace-qualifiers to fix "changes meaning of name" warnings.  - ./qpid/ptr_map.h:51: fixed "qualified return value" warning.  - use const char* for "conversion from string constant to ‘char*’" warnings
Applied patch from https://issues.apache.org/jira/browse/QPID-869
remove depenency on boost/date_time, causes warnings with gcc 4.3.


git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid@640806 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/sys/Socket.h
@@ -93,8 +93,8 @@ public:
      */
     std::string getLocalAddress() const;
 
-    uint getLocalPort() const;
-    uint getRemotePort() const;
+    uint16_t getLocalPort() const;
+    uint16_t getRemotePort() const;
 
     
     /** Accept a connection from a socket that is already listening
